==> 8. Comparison of measures.rmd <==
---
title: "8. Comparison of measures"
author: "Shy-Ann Moehamatdjalil"
date: "27-5-2021"
output: html_document
---

# DEPENDS ON: Data/comment_overview_CLEAN.xlsx, Data/Prediction files/1.NRC Prediction.xlsx, Data/Prediction files/2.DANEW Prediction.xlsx,
#             Data/Prediction files/3. Pattern Prediction.csv, Data/Prediction files/4. Polyglot Prediction.csv, 
#             Data/Prediction files/5. Support Vector Machine Prediction.xlsx, Data/Prediction files/6.Naive Bayes Prediction.xlsx

#             Data/Extra analyse/Prediction files/1.NRC Prediction_ZT.xlsx, Data/Extra analyse/Prediction files/2.DANEW Prediction_ZT.xlsx,
#             Data/Extra analyse/Prediction files/3. Pattern Prediction_ZT.csv, Data/Extra analyse/Prediction files/4. Polyglot
#             Prediction_ZT.csv, Data/Extra analyse/Prediction files/5. Support Vector Machine Prediction_ZT.xlsx, Data/Extra analyse/Prediction
#             files/6. Naive Bayes Prediction_ZT.xlsx

## Load packages

```{r message=FALSE, warning=FALSE}
library(tidyverse)
library(tidytext)
library(readxl)
library(writexl)
```

### Load all the prediction files in 
```{r message=FALSE, warning=FALSE}
comments <- read_excel("./Data/comment_overview_CLEAN.xlsx", col_types = "text")
NRC <- read_excel("./Data/Prediction files/1.NRC Prediction.xlsx")
DANEW <- read_excel("./Data/Prediction files/2.DANEW Prediction.xlsx")
Pattern <- read_csv("./Data/Prediction files/3. Pattern Prediction.csv")
Polyglot <- read_csv("./Data/Prediction files/4. Polyglot Prediction.csv")
SVM <- read_excel("./Data/Prediction files/5. Support Vector Machine Prediction.xlsx")
NB <- read_excel("./Data/Prediction files/6. Naive Bayes Prediction.xlsx")
```

### Make tables
Make functions for calculating accuracy, precision, recall and F-measure
```{r}
accuracy <- function(proportions){
  acc = prop.table(proportions) %>% 
    diag() %>% 
    sum() %>% 
    round(2)
  return(acc)
}

precision <- function(table, class){
  
  if (class == "Negative"){
    TP <- table["Negative", "Negative"]
    FP <- table["Neutral", "Negative"] + table["Positive", "Negative"]
  } else if(class == "Neutral"){
    TP <- table["Neutral", "Neutral"]
    FP <- table["Negative", "Neutral"] + table["Positive", "Neutral"]
  } else {
    TP <- table["Positive", "Positive"]
    FP <- table["Negative", "Positive"] + table["Neutral", "Positive"]
  }
  
  precision = TP/(TP + FP)
  precision = round(precision, 2)
  return(precision)

}
  
recall <- function(table, class){
  
  if (class == "Negative"){
    TP <- table["Negative", "Negative"]
    FN <- table["Negative", "Neutral"] + table["Negative", "Positive"]
  } else if(class == "Neutral"){
    TP <- table["Neutral", "Neutral"]
    FN <- table["Neutral", "Negative"] + table["Neutral", "Positive"]
  } else {
    TP <- table["Positive", "Positive"]
    FN <- table["Positive", "Negative"] + table["Positive", "Neutral"]
  }
  
  recall = TP/(TP + FN)
  recall = round(recall, 2)
  return(recall)

}

F1 <- function(precision, recall){
  F1_score = 2 * (precision * recall)/(precision + recall)
  F1_score = round(F1_score, 2)
  return(F1_score)
}

```


NRC
```{r}
# Make a frequency table that compares manual and automated
NRC_table <- table(NRC$ma_sentiment, NRC$sentiment_auto)
names(dimnames(NRC_table)) <- c("MANUAL", "AUTOMATED")

NRC_table

# Calculate accuracy
NRC_acc <- accuracy(NRC_table)

# For Negative class calculate precision, recall, f1
NRC_N_pre <- precision(NRC_table, "Negative")
NRC_N_rec <- recall(NRC_table, "Negative")
NRC_N_F1 <- F1(NRC_N_pre, NRC_N_rec)

# For Neutral class calculate precision, recall, f1
NRC_Neu_pre <- precision(NRC_table, "Neutral")
NRC_Neu_rec <- recall(NRC_table, "Neutral")
NRC_Neu_F1 <- F1(NRC_Neu_pre, NRC_Neu_rec)

# For Positive class calculate precision, recall, f1
NRC_P_pre <- precision(NRC_table, "Positive")
NRC_P_rec <- recall(NRC_table, "Positive")
NRC_P_F1 <- F1(NRC_P_pre, NRC_P_rec)

# Calculate macro F1
NRC_MAF1 <- mean(c(NRC_N_F1, NRC_Neu_F1, NRC_P_F1)) %>% round(2)

# Store all values in a list
NRC_values <- list("Method" = "NRC", "Acc" = NRC_acc,"MAF1" = NRC_MAF1,
                   "Neg_Precision" = NRC_N_pre, "Neg_Recall" = NRC_N_rec, "Neg_F1" = NRC_N_F1,
                   "Neu_Precision" = NRC_Neu_pre, "Neu_Recall" = NRC_Neu_rec, "Neu_F1" = NRC_Neu_F1,
                   "Pos_Precision" = NRC_P_pre, "Pos_Recall" = NRC_P_rec, "Pos_F1" = NRC_P_F1)

```

DANEW
```{r}
# Make a frequency table that compares manual and automated
DANEW_table <- table(DANEW$ma_sentiment, DANEW$sentiment_auto)
names(dimnames(DANEW_table)) <- c("MANUAL", "AUTOMATED")

DANEW_table

# Calculate accuracy
DANEW_acc <- accuracy(DANEW_table)

# For Negative class calculate precision, recall, f1
DANEW_N_pre <- precision(DANEW_table, "Negative")
DANEW_N_rec <- recall(DANEW_table, "Negative")
DANEW_N_F1 <- F1(DANEW_N_pre, DANEW_N_rec)

# For Neutral class calculate precision, recall, f1
DANEW_Neu_pre <- precision(DANEW_table, "Neutral")
DANEW_Neu_rec <- recall(DANEW_table, "Neutral")
DANEW_Neu_F1 <- F1(DANEW_Neu_pre, DANEW_Neu_rec)

# For Positive class calculate precision, recall, f1
DANEW_P_pre <- precision(DANEW_table, "Positive")
DANEW_P_rec <- recall(DANEW_table, "Positive")
DANEW_P_F1 <- F1(DANEW_P_pre, DANEW_P_rec)

# Calculate macro F1
DANEW_MAF1 <- mean(c(DANEW_N_F1, DANEW_Neu_F1, DANEW_P_F1)) %>% round(2)


# Store all values in a list
DANEW_values <- list("Method" = "DANEW", "Acc" = DANEW_acc,"MAF1" = DANEW_MAF1,
                     "Neg_Precision" = DANEW_N_pre, "Neg_Recall" = DANEW_N_rec, "Neg_F1" = DANEW_N_F1,
                     "Neu_Precision" = DANEW_Neu_pre, "Neu_Recall" = DANEW_Neu_rec, "Neu_F1" = DANEW_Neu_F1,
                     "Pos_Precision" = DANEW_P_pre, "Pos_Recall" = DANEW_P_rec, "Pos_F1" = DANEW_P_F1)
```

Pattern
```{r}
# Make a frequency table that compares manual and automated
Pattern_table <- table(Pattern$ma_sentiment, Pattern$sentiment_auto)
names(dimnames(Pattern_table)) <- c("MANUAL", "AUTOMATED")

Pattern_table

# Calculate accuracy
Pattern_acc <- accuracy(Pattern_table)

# For Negative class calculate precision, recall, f1
Pattern_N_pre <- precision(Pattern_table, "Negative")
Pattern_N_rec <- recall(Pattern_table, "Negative")
Pattern_N_F1 <- F1(Pattern_N_pre, Pattern_N_rec)

# For Neutral class calculate precision, recall, f1
Pattern_Neu_pre <- precision(Pattern_table, "Neutral")
Pattern_Neu_rec <- recall(Pattern_table, "Neutral")
Pattern_Neu_F1 <- F1(Pattern_Neu_pre, Pattern_Neu_rec)

# For Positive class calculate precision, recall, f1
Pattern_P_pre <- precision(Pattern_table, "Positive")
Pattern_P_rec <- recall(Pattern_table, "Positive")
Pattern_P_F1 <- F1(Pattern_P_pre, Pattern_P_rec)

# Calculate macro F1
Pattern_MAF1 <- mean(c(Pattern_N_F1, Pattern_Neu_F1, Pattern_P_F1)) %>% round(2)

# Store all values in a list
Pattern_values <- list("Method" = "Pattern", "Acc" = Pattern_acc, "MAF1" = Pattern_MAF1,
                       "Neg_Precision" = Pattern_N_pre, "Neg_Recall" = Pattern_N_rec, "Neg_F1" = Pattern_N_F1,
                       "Neu_Precision" = Pattern_Neu_pre, "Neu_Recall" = Pattern_Neu_rec, "Neu_F1" = Pattern_Neu_F1,
                       "Pos_Precision" = Pattern_P_pre, "Pos_Recall" = Pattern_P_rec, "Pos_F1" = Pattern_P_F1)
```

Polyglot
```{r}
# Make a frequency table that compares manual and automated
Polyglot_table <- table(Polyglot$ma_sentiment, Polyglot$sentiment_auto)
names(dimnames(Polyglot_table)) <- c("MANUAL", "AUTOMATED")

Polyglot_table

# Calculate accuracy
Polyglot_acc <- accuracy(Polyglot_table)

# For Negative class calculate precision, recall, f1
Polyglot_N_pre <- precision(Polyglot_table, "Negative")
Polyglot_N_rec <- recall(Polyglot_table, "Negative")
Polyglot_N_F1 <- F1(Polyglot_N_pre, Polyglot_N_rec)

# For Neutral class calculate precision, recall, f1
Polyglot_Neu_pre <- precision(Polyglot_table, "Neutral")
Polyglot_Neu_rec <- recall(Polyglot_table, "Neutral")
Polyglot_Neu_F1 <- F1(Polyglot_Neu_pre, Polyglot_Neu_rec)

# For Positive class calculate precision, recall, f1
Polyglot_P_pre <- precision(Polyglot_table, "Positive")
Polyglot_P_rec <- recall(Polyglot_table, "Positive")
Polyglot_P_F1 <- F1(Polyglot_P_pre, Polyglot_P_rec)

# Calculate macro F1
Polyglot_MAF1 <- mean(c(Polyglot_N_F1, Polyglot_Neu_F1, Polyglot_P_F1)) %>% round(2)

# Store all values in a list
Polyglot_values <- list("Method" = "Polyglot", "Acc" = Polyglot_acc, "MAF1" = Polyglot_MAF1,
                        "Neg_Precision" = Polyglot_N_pre, "Neg_Recall" = Polyglot_N_rec, "Neg_F1" = Polyglot_N_F1,
                        "Neu_Precision" = Polyglot_Neu_pre, "Neu_Recall" = Polyglot_Neu_rec, "Neu_F1" = Polyglot_Neu_F1,
                        "Pos_Precision" = Polyglot_P_pre, "Pos_Recall" = Polyglot_P_rec, "Pos_F1" = Polyglot_P_F1)
```

SVM
```{r}
# Make a frequency table that compares manual and automated
SVM_table <- table(SVM$ma_sentiment, SVM$predicted)
names(dimnames(SVM_table)) <- c("MANUAL", "AUTOMATED")

SVM_table

# Calculate accuracy
SVM_acc <- accuracy(SVM_table)

# For Negative class calculate precision, recall, f1
SVM_N_pre <- precision(SVM_table, "Negative")
SVM_N_rec <- recall(SVM_table, "Negative")
SVM_N_F1 <- F1(SVM_N_pre, SVM_N_rec)

# For Neutral class calculate precision, recall, f1
SVM_Neu_pre <- precision(SVM_table, "Neutral")
SVM_Neu_rec <- recall(SVM_table, "Neutral")
SVM_Neu_F1 <- F1(SVM_Neu_pre, SVM_Neu_rec)

# For Positive class calculate precision, recall, f1
SVM_P_pre <- precision(SVM_table, "Positive")
SVM_P_rec <- recall(SVM_table, "Positive")
SVM_P_F1 <- F1(SVM_P_pre, SVM_P_rec)

# Calculate macro F1
SVM_MAF1 <- mean(c(SVM_N_F1, SVM_Neu_F1, SVM_P_F1)) %>% round(2)

# Store all values in a list
SVM_values <- list("Method" = "SVM", "Acc" = SVM_acc, "MAF1" = SVM_MAF1,
                   "Neg_Precision" = SVM_N_pre, "Neg_Recall" = SVM_N_rec, "Neg_F1" = SVM_N_F1,
                   "Neu_Precision" = SVM_Neu_pre, "Neu_Recall" = SVM_Neu_rec, "Neu_F1" = SVM_Neu_F1,
                   "Pos_Precision" = SVM_P_pre, "Pos_Recall" = SVM_P_rec, "Pos_F1" = SVM_P_F1)

```

NB
```{r}
# Make a frequency table that compares manual and automated
NB_table <- table(NB$ma_sentiment, NB$predicted)
names(dimnames(NB_table)) <- c("MANUAL", "AUTOMATED")

NB_table

# Calculate accuracy
NB_acc <- accuracy(NB_table)

# For Negative class calculate precision, recall, f1
NB_N_pre <- precision(NB_table, "Negative")
NB_N_rec <- recall(NB_table, "Negative")
NB_N_F1 <- F1(NB_N_pre, NB_N_rec)

# For Neutral class calculate precision, recall, f1
NB_Neu_pre <- precision(NB_table, "Neutral")
NB_Neu_rec <- recall(NB_table, "Neutral")
NB_Neu_F1 <- F1(NB_Neu_pre, NB_Neu_rec)

# For Positive class calculate precision, recall, f1
NB_P_pre <- precision(NB_table, "Positive")
NB_P_rec <- recall(NB_table, "Positive")
NB_P_F1 <- F1(NB_P_pre, NB_P_rec)

# Calculate macro F1
NB_MAF1 <- mean(c(NB_N_F1, NB_Neu_F1, NB_P_F1)) %>% round(2)

# Store all values in a list
NB_values <- list( "Method" = "NB", "Acc" = NB_acc, "MAF1" = NB_MAF1,
                   "Neg_Precision" = NB_N_pre, "Neg_Recall" = NB_N_rec, "Neg_F1" = NB_N_F1,
                   "Neu_Precision" = NB_Neu_pre, "Neu_Recall" = NB_Neu_rec, "Neu_F1" = NB_Neu_F1,
                   "Pos_Precision" = NB_P_pre, "Pos_Recall" = NB_P_rec, "Pos_F1" = NB_P_F1)

```

### Combine all lists into one dataframe
```{r}
comparison_df <- rbind(data.frame(NRC_values), data.frame(DANEW_values), data.frame(Pattern_values),
                       data.frame(Polyglot_values), data.frame(SVM_values),data.frame(NB_values))
comparison_df
```


### ------------------- EXTRA ANALYSIS --------- ####

### Load all the prediction files in 
```{r message=FALSE, warning=FALSE}
NRC_ZT <- read_excel("./Data/Extra analyse/Prediction files/1.NRC Prediction_ZT.xlsx")
DANEW_ZT <- read_excel("./Data/Extra analyse/Prediction files/2.DANEW Prediction_ZT.xlsx")
Pattern_ZT <- read_csv("./Data/Extra analyse/Prediction files/3. Pattern Prediction_ZT.csv")
Polyglot_ZT <- read_csv("./Data/Extra analyse/Prediction files/4. Polyglot Prediction_ZT.csv")
SVM_ZT <- read_excel("./Data/Extra analyse/Prediction files/5. Support Vector Machine Prediction_ZT.xlsx")
NB_ZT <- read_excel("./Data/Extra analyse/Prediction files/6. Naive Bayes Prediction_ZT.xlsx")
```

NRC
```{r}
# Make a frequency table that compares manual and automated
NRC_ZT_table <- table(NRC_ZT$ma_sentiment, NRC_ZT$sentiment_auto)
names(dimnames(NRC_ZT_table)) <- c("MANUAL", "AUTOMATED")

NRC_ZT_table

# Calculate accuracy
NRC_ZT_acc <- accuracy(NRC_ZT_table)

# For Negative class calculate precision, recall, f1
NRC_ZT_N_pre <- precision(NRC_ZT_table, "Negative")
NRC_ZT_N_rec <- recall(NRC_ZT_table, "Negative")
NRC_ZT_N_F1 <- F1(NRC_ZT_N_pre, NRC_ZT_N_rec)

# For Neutral class calculate precision, recall, f1
NRC_ZT_Neu_pre <- precision(NRC_ZT_table, "Neutral")
NRC_ZT_Neu_rec <- recall(NRC_ZT_table, "Neutral")
NRC_ZT_Neu_F1 <- F1(NRC_ZT_Neu_pre, NRC_ZT_Neu_rec)

# For Positive class calculate precision, recall, f1
NRC_ZT_P_pre <- precision(NRC_ZT_table, "Positive")
NRC_ZT_P_rec <- recall(NRC_ZT_table, "Positive")
NRC_ZT_P_F1 <- F1(NRC_ZT_P_pre, NRC_ZT_P_rec)

# Calculate macro F1
NRC_ZT_MAF1 <- mean(c(NRC_ZT_N_F1, NRC_ZT_Neu_F1, NRC_ZT_P_F1)) %>% round(2)

# Store all values in a list
NRC_ZT_values <- list("Method" = "NRC_ZT", "Acc" = NRC_ZT_acc, "MAF1" = NRC_ZT_MAF1,
                   "Neg_Precision" = NRC_ZT_N_pre, "Neg_Recall" = NRC_ZT_N_rec, "Neg_F1" = NRC_ZT_N_F1,
                   "Neu_Precision" = NRC_ZT_Neu_pre, "Neu_Recall" = NRC_ZT_Neu_rec, "Neu_F1" = NRC_ZT_Neu_F1,
                   "Pos_Precision" = NRC_ZT_P_pre, "Pos_Recall" = NRC_ZT_P_rec, "Pos_F1" = NRC_ZT_P_F1)

```

DANEW
```{r}
# Make a frequency table that compares manual and automated
DANEW_ZT_table <- table(DANEW_ZT$ma_sentiment, DANEW_ZT$sentiment_auto)
names(dimnames(DANEW_ZT_table)) <- c("MANUAL", "AUTOMATED")

DANEW_ZT_table

# Calculate accuracy
DANEW_ZT_acc <- accuracy(DANEW_ZT_table)

# For Negative class calculate precision, recall, f1
DANEW_ZT_N_pre <- precision(DANEW_ZT_table, "Negative")
DANEW_ZT_N_rec <- recall(DANEW_ZT_table, "Negative")
DANEW_ZT_N_F1 <- F1(DANEW_ZT_N_pre, DANEW_ZT_N_rec)

# For Neutral class calculate precision, recall, f1
DANEW_ZT_Neu_pre <- precision(DANEW_ZT_table, "Neutral")
DANEW_ZT_Neu_rec <- recall(DANEW_ZT_table, "Neutral")
DANEW_ZT_Neu_F1 <- F1(DANEW_ZT_Neu_pre, DANEW_ZT_Neu_rec)

# For Positive class calculate precision, recall, f1
DANEW_ZT_P_pre <- precision(DANEW_ZT_table, "Positive")
DANEW_ZT_P_rec <- recall(DANEW_ZT_table, "Positive")
DANEW_ZT_P_F1 <- F1(DANEW_ZT_P_pre, DANEW_ZT_P_rec)

# Calculate macro F1
DANEW_ZT_MAF1 <- mean(c(DANEW_ZT_N_F1, DANEW_ZT_Neu_F1, DANEW_ZT_P_F1)) %>% round(2)

# Store all values in a list
DANEW_ZT_values <- list("Method" = "DANEW_ZT", "Acc" = DANEW_ZT_acc, "MAF1" = DANEW_ZT_MAF1,
                        "Neg_Precision" = DANEW_ZT_N_pre, "Neg_Recall" = DANEW_ZT_N_rec, "Neg_F1" = DANEW_ZT_N_F1,
                        "Neu_Precision" = DANEW_ZT_Neu_pre,"Neu_Recall" = DANEW_ZT_Neu_rec,
                        "Neu_F1"= DANEW_ZT_Neu_F1,
                        "Pos_Precision" = DANEW_ZT_P_pre, "Pos_Recall" = DANEW_ZT_P_rec, "Pos_F1" = DANEW_ZT_P_F1)
```

Pattern
```{r}
# Make a frequency table that compares manual and automated
Pattern_ZT_table <- table(Pattern_ZT$ma_sentiment, Pattern_ZT$sentiment_auto)
names(dimnames(Pattern_ZT_table)) <- c("MANUAL", "AUTOMATED")

Pattern_ZT_table

# Calculate accuracy
Pattern_ZT_acc <- accuracy(Pattern_ZT_table)

# For Negative class calculate precision, recall, f1
Pattern_ZT_N_pre <- precision(Pattern_ZT_table, "Negative")
Pattern_ZT_N_rec <- recall(Pattern_ZT_table, "Negative")
Pattern_ZT_N_F1 <- F1(Pattern_ZT_N_pre, Pattern_ZT_N_rec)

# For Neutral class calculate precision, recall, f1
Pattern_ZT_Neu_pre <- precision(Pattern_ZT_table, "Neutral")
Pattern_ZT_Neu_rec <- recall(Pattern_ZT_table, "Neutral")
Pattern_ZT_Neu_F1 <- F1(Pattern_ZT_Neu_pre, Pattern_ZT_Neu_rec)

# For Positive class calculate precision, recall, f1
Pattern_ZT_P_pre <- precision(Pattern_ZT_table, "Positive")
Pattern_ZT_P_rec <- recall(Pattern_ZT_table, "Positive")
Pattern_ZT_P_F1 <- F1(Pattern_ZT_P_pre, Pattern_ZT_P_rec)

# Calculate macro F1
Pattern_ZT_MAF1 <- mean(c(Pattern_ZT_N_F1, Pattern_ZT_Neu_F1, Pattern_ZT_P_F1)) %>% round(2)

# Store all values in a list
Pattern_ZT_values <- list("Method" = "Pattern_ZT", "Acc" = Pattern_ZT_acc, "MAF1" = Pattern_ZT_MAF1,
                       "Neg_Precision" = Pattern_ZT_N_pre, "Neg_Recall" = Pattern_ZT_N_rec, 
                       "Neg_F1" = Pattern_ZT_N_F1,
                       "Neu_Precision" = Pattern_ZT_Neu_pre, "Neu_Recall" = Pattern_ZT_Neu_rec, 
                       "Neu_F1" = Pattern_ZT_Neu_F1,
                       "Pos_Precision" = Pattern_ZT_P_pre, "Pos_Recall" = Pattern_ZT_P_rec, 
                       "Pos_F1" = Pattern_ZT_P_F1)
```

Polyglot
```{r}
# Make a frequency table that compares manual and automated
Polyglot_ZT_table <- table(Polyglot_ZT$ma_sentiment, Polyglot_ZT$sentiment_auto)
names(dimnames(Polyglot_ZT_table)) <- c("MANUAL", "AUTOMATED")

Polyglot_ZT_table

# Calculate accuracy
Polyglot_ZT_acc <- accuracy(Polyglot_ZT_table)

# For Negative class calculate precision, recall, f1
Polyglot_ZT_N_pre <- precision(Polyglot_ZT_table, "Negative")
Polyglot_ZT_N_rec <- recall(Polyglot_ZT_table, "Negative")
Polyglot_ZT_N_F1 <- F1(Polyglot_ZT_N_pre, Polyglot_ZT_N_rec)

# For Neutral class calculate precision, recall, f1
Polyglot_ZT_Neu_pre <- precision(Polyglot_ZT_table, "Neutral")
Polyglot_ZT_Neu_rec <- recall(Polyglot_ZT_table, "Neutral")
Polyglot_ZT_Neu_F1 <- F1(Polyglot_ZT_Neu_pre, Polyglot_ZT_Neu_rec)

# For Positive class calculate precision, recall, f1
Polyglot_ZT_P_pre <- precision(Polyglot_ZT_table, "Positive")
Polyglot_ZT_P_rec <- recall(Polyglot_ZT_table, "Positive")
Polyglot_ZT_P_F1 <- F1(Polyglot_ZT_P_pre, Polyglot_ZT_P_rec)

# Calculate macro F1
Polyglot_ZT_MAF1 <- mean(c(Polyglot_ZT_N_F1, Polyglot_ZT_Neu_F1, Polyglot_ZT_P_F1)) %>% round(2)

# Store all values in a list
Polyglot_ZT_values <- list("Method" = "Polyglot_ZT", "Acc" = Polyglot_ZT_acc, "MAF1" = Polyglot_ZT_MAF1,
                        "Neg_Precision" = Polyglot_ZT_N_pre, "Neg_Recall" = Polyglot_ZT_N_rec, 
                        "Neg_F1" = Polyglot_ZT_N_F1,
                        "Neu_Precision" = Polyglot_ZT_Neu_pre, "Neu_Recall" = Polyglot_ZT_Neu_rec,
                        "Neu_F1" = Polyglot_ZT_Neu_F1,
                        "Pos_Precision" = Polyglot_ZT_P_pre, "Pos_Recall" = Polyglot_ZT_P_rec, 
                        "Pos_F1" = Polyglot_ZT_P_F1)
```

SVM
```{r}
# Make a frequency table that compares manual and automated
SVM_ZT_table <- table(SVM_ZT$ma_sentiment, SVM_ZT$predicted)
names(dimnames(SVM_ZT_table)) <- c("MANUAL", "AUTOMATED")

SVM_ZT_table

# Calculate accuracy
SVM_ZT_acc <- accuracy(SVM_ZT_table)

# For Negative class calculate precision, recall, f1
SVM_ZT_N_pre <- precision(SVM_ZT_table, "Negative")
SVM_ZT_N_rec <- recall(SVM_ZT_table, "Negative")
SVM_ZT_N_F1 <- F1(SVM_ZT_N_pre, SVM_ZT_N_rec)

# For Neutral class calculate precision, recall, f1
SVM_ZT_Neu_pre <- precision(SVM_ZT_table, "Neutral")
SVM_ZT_Neu_rec <- recall(SVM_ZT_table, "Neutral")
SVM_ZT_Neu_F1 <- F1(SVM_ZT_Neu_pre, SVM_ZT_Neu_rec)

# For Positive class calculate precision, recall, f1
SVM_ZT_P_pre <- precision(SVM_ZT_table, "Positive")
SVM_ZT_P_rec <- recall(SVM_ZT_table, "Positive")
SVM_ZT_P_F1 <- F1(SVM_ZT_P_pre, SVM_ZT_P_rec)

# Calculate macro F1
SVM_ZT_MAF1 <- mean(c(SVM_ZT_N_F1, SVM_ZT_Neu_F1, SVM_ZT_P_F1)) %>% round(2)

# Store all values in a list
SVM_ZT_values <- list("Method" = "SVM_ZT", "Acc" = SVM_ZT_acc, "MAF1" = SVM_ZT_MAF1,
                   "Neg_Precision" = SVM_ZT_N_pre, "Neg_Recall" = SVM_ZT_N_rec, "Neg_F1" = SVM_ZT_N_F1,
                   "Neu_Precision" = SVM_ZT_Neu_pre, "Neu_Recall" = SVM_ZT_Neu_rec, "Neu_F1" = SVM_ZT_Neu_F1,
                   "Pos_Precision" = SVM_ZT_P_pre, "Pos_Recall" = SVM_ZT_P_rec, "Pos_F1" = SVM_ZT_P_F1)

```


NB
```{r}
# Make a frequency table that compares manual and automated
NB_ZT_table <- table(NB_ZT$ma_sentiment, NB_ZT$predicted)
names(dimnames(NB_ZT_table)) <- c("MANUAL", "AUTOMATED")

NB_ZT_table

# Calculate accuracy
NB_ZT_acc <- accuracy(NB_ZT_table)

# For Negative class calculate precision, recall, f1
NB_ZT_N_pre <- precision(NB_ZT_table, "Negative")
NB_ZT_N_rec <- recall(NB_ZT_table, "Negative")
NB_ZT_N_F1 <- F1(NB_ZT_N_pre, NB_ZT_N_rec)

# For Neutral class calculate precision, recall, f1
NB_ZT_Neu_pre <- precision(NB_ZT_table, "Neutral")
NB_ZT_Neu_rec <- recall(NB_ZT_table, "Neutral")
NB_ZT_Neu_F1 <- F1(NB_ZT_Neu_pre, NB_ZT_Neu_rec)

# For Positive class calculate precision, recall, f1
NB_ZT_P_pre <- precision(NB_ZT_table, "Positive")
NB_ZT_P_rec <- recall(NB_ZT_table, "Positive")
NB_ZT_P_F1 <- F1(NB_ZT_P_pre, NB_ZT_P_rec)

# Calculate macro F1
NB_ZT_MAF1 <- mean(c(NB_ZT_N_F1, NB_ZT_Neu_F1, NB_ZT_P_F1)) %>% round(2)

# Store all values in a list
NB_ZT_values <- list( "Method" = "NB_ZT", "Acc" = NB_ZT_acc, "MAF1" = NB_ZT_MAF1,
                   "Neg_Precision" = NB_ZT_N_pre, "Neg_Recall" = NB_ZT_N_rec, "Neg_F1" = NB_ZT_N_F1,
                   "Neu_Precision" = NB_ZT_Neu_pre, "Neu_Recall" = NB_ZT_Neu_rec, "Neu_F1" = NB_ZT_Neu_F1,
                   "Pos_Precision" = NB_ZT_P_pre, "Pos_Recall" = NB_ZT_P_rec, "Pos_F1" = NB_ZT_P_F1)

```

### Combine all lists into one dataframe
```{r}
comparison_df_ZT <- rbind(data.frame(NRC_ZT_values), data.frame(DANEW_ZT_values), data.frame(Pattern_ZT_values),
                       data.frame(Polyglot_ZT_values), data.frame(SVM_ZT_values), data.frame(NB_ZT_values))
comparison_df_ZT
```
